Moving greatwyrm inventory and garden of forking paths session pages.
